Let .deb to check library dependancy, so it won't need code to install dependencies.
Indeed when you install this package with this commit, you'll just need to apply
'apt-get install -f' to do this application ready to read PDFs.

It's good for QtMoko, because Raptor will automatically apply this command
at the end of installation and so the application will be ready directly.
And so it fixes really well the issue due to the new multi-arch of Debian/Wheezy
and such issues due to future development of Debian.